Updating www based on experience working in Coinbase repo. Updating mock file structure for easier usage with websocket.
