+++ .gitignore
@@ -9,11 +9,5 @@ credentials.json
 action_history*.csv
 position_history*.csv
 
-BCHUSD_candlesticks_from_1672531500_to_1704067080.csv
-BCHUSD_ticks_from_1672531500_to_1704067080.csv
-
-BTCUSD_candlesticks_from_1672531500_to_1688083080.csv
-BTCUSD_ticks_from_1672531500_to_1688083080.csv
-
-SOLUSD_candlesticks_from_1672531500_to_1704067080.csv
-SOLUSD_ticks_from_1672531500_to_1704067080.csv
+ticks*
+candlesticks*
