+++ mock/sim_candlesticks_SOLUSD_from_1704067500_to_1704585480.csv
