--- www/.gitignore
@@ -1,4 +1,0 @@
-node_modules
-dist
-.parcel-cache
-package-lock.json
