+++ www/index.html
@@ -5,7 +5,7 @@
   </head>
   <body>
     <div style="width: 800px;"><canvas id="my_chart"></canvas></div>
-    <script type="module" src="my_chart.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+    <script src="messenger.js"></script>
   </body>
 </html>
-
