--- www/package.json
@@ -1,17 +1,0 @@
-{
-    "name": "Trade Bot",
-    "version": "1.0.0",
-    "license": "MIT",
-    "scripts": {
-      "dev": "parcel src/index.html",
-      "build": "parcel build src/index.html"
-    },
-    "devDependencies": {
-      "parcel": "^2.6.2"
-    },
-    "dependencies": {
-      "@cubejs-client/core": "^0.31.0",
-      "@cubejs-client/ws-transport": "0.34.60",
-      "chart.js": "^4.0.0"
-    }
-  }
