+++ www/show_message.js
@@ -1,0 +1,12 @@
+window.addEventListener("DOMContentLoaded", () => {
+    const messages = document.createElement("ul");
+    document.body.appendChild(messages);
+  
+    const websocket = new WebSocket("ws://localhost:5678/");
+    websocket.onmessage = ({ data }) => {
+      const message = document.createElement("li");
+      const content = document.createTextNode(data);
+      message.appendChild(content);
+      messages.appendChild(message);
+    };
+  });
